+++ resources/views/importUtm.blade.php
@@ -39,7 +39,7 @@
         </div>
     </div>
 </form>
-<script src="https://code.jquery.com/jquery-3.4.1.min.js"></script>
+<script src="{{asset('/js/jquery-3.4.1.min.js')}}"></script>
 <script type="text/javascript">
     $(document).ready(function () {
         $("#linkDwonload").hide();
